[python][UHI] Add plotting tests for TH2 and TH3
